+++ fxn/types/predictor.py
@@ -25,7 +25,6 @@ class AccessMode (str, Enum):
     """
     Public = "PUBLIC"
     Private = "PRIVATE"
-    Protected = "PROTECTED"
 
 class PredictorType (str, Enum):
     """